Implementation of JWT

+++ resources/item.py
@@ -1,29 +1,31 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from schemas import ItemSchema, ItemUpdateSchema
 from models import ItemModel
 from sqlalchemy.exc import SQLAlchemyError
 from db import db
-
+from flask_jwt_extended import jwt_required, get_jwt
 blp = Blueprint("Items", "__name__", description = "Operations on items")
 
 
-@blp.route("/items/<string:item_id>")
+@blp.route("/items/<int:item_id>")
 class Item(MethodView):
-
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get (self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
-    
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
-        return {"message": "Item Deleted"}
+        return {"message": "Item deleted."}
 
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
@@ -46,10 +48,12 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
